Add recipes and img folder to root
Create two new folders (img and recipes) to root of the webpage.

The img folder is meant to store images for general use on pages.

The recipes folder is meant to store all recipe related webpages.

Also create a HTML page for a lasagna recipe that includes the
ingredients and steps to make the dish.

The index.html now contains a link to the lasagna recipe page for
redirecting.